requirements.txt
beautifulsoup4==4.9.1
blinker==1.4
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
google-api-core==1.16.0
google-auth==1.19.1
google-cloud-core==1.3.0
google-cloud-datastore==1.13.1
googleapis-common-protos==1.52.0
grpcio==1.27.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
yagmail==0.11.224
